src/main/java/com/blank04/universitycms/service/TeacherService.java
package com.blank04.universitycms.service;

import com.blank04.universitycms.model.user.impl.Teacher;
import com.blank04.universitycms.repository.TeacherRepository;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.sql.SQLException;
import java.util.List;
import java.util.Optional;

@Service
public class TeacherService {
    private final TeacherRepository teacherRepository;

    public TeacherService(TeacherRepository teacherRepository) {
        this.teacherRepository = teacherRepository;
    }

    @Transactional
    public Teacher save(Teacher teacher) {
        return teacherRepository.save(teacher);
    }

    @Transactional(readOnly = true)
    public Optional<Teacher> findById(Long id) {
        return teacherRepository.findById(id);
    }

    @Transactional(readOnly = true)
    public List<Teacher> findAll() {
        return teacherRepository.findAll();
    }

    @Transactional
    public void deleteById(Long id) throws SQLException {
        Optional<Teacher> entity = findById(id);
        if (entity.isPresent()) {
            teacherRepository.deleteById(id);
        } else throw new SQLException("Teacher with id="+ id + " doesn't exist");
    }
}
